fix(app-server): address code-review findings
Correctness:
- Route app-server chat submissions back to the originating channel.
  submitOp now records session ownership so a dedicated connection
  receives its own event stream instead of the stream leaking to the UI
  channel (and the external client no longer hangs with no events).
- Stop falling back to the primary session for a dedicated channel whose
  session has vanished — fail loudly rather than execute the request
  against the user's live UI conversation.
- Return the submission id from submitOp so chat.send surfaces a stable
  runId (was always undefined) for automation.
- Bypass the rate limiter (not just the queue) for health/readiness so
  liveness probes stay observable under load.
- Share a single eventMsgToName wire-name mapping between ServerChannel
  and AppServerChannel so the two cannot drift and leak unmapped events
  across scopes.

Cleanup:
- Remove the write-only isLoopback connection field (the real gate is the
  bind-time loopback/allowLan check + verifyClient origin rejection).
- Collapse the dead 'disabled' ? 'disabled' : 'disabled' status ternary.
- Extract buildAvailableEvents into @workx/ws-server, used by both the
  headless handshake and the app-server processor.

Validation: tsc --noEmit clean; vitest app-server (64) + server (329) +
ws-server + core/channels + desktop-runtime bootstrap all green.
